More script work for clean startup and shutdown.

+++ src/main/bin/kill.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
 pid=$(cat ${THERMOSTAT_HOME}/pid)
-echo "killing pid:" ${pid}
-kill -9 ${pid}
+if [ -n "$pid" ]; then
+    echo " "
+    echo "killing pid:" ${pid}
+    kill -9 ${pid}
+    if [ $? -eq 0 ]; then
+        rm ${THERMOSTAT_HOME}/pid
+    fi
+fi
 
 #pid=`ps xauww | grep java | grep ControlThread | cut -d" " -f3`
 #if [ -n "$pid" ]; then
